Part 4 and some cleanup
We now have statistic tracking and encountered an interesting possible
language bug, at least something that wasn't happening as I expected it
to and was resulting in significantly... wonky... timing statistics for
the hash function. I have commented in there what it likely is and also
added a quick test to check some things, though I am planning on
removing them as they aren't strictly speaking a part of this
assignment.