Attempted to use go for PRT... it has issues
For some reason, setting the body of an HTTP request in go seems...
error prone at best. In every case I was unable to actually receive the
body of the request (via POST) on the server, though it was sending a
content length value that was non-zero. In trying multiple ways, I have
determined that it isn't worth the time to fuss over the http requests
not working properly, and instead I am going to fall back to standard
TCP sockets.